simplify upload and metadata
